requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
python-dotenv
pymongo
flask-pymongo
pytesseract
bcrypt
flask-socketio  
eventlet